# najah-delivery-os/services/express-api/openapi.yaml
openapi: 3.0.3
info:
  title: Najah Delivery API
  description: Express API service for Najah Delivery OS - providing comprehensive delivery management endpoints
  version: 1.0.0
  contact:
    name: Najah Delivery Support
    email: support@najahdelivery.com
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.najahdelivery.com/v1
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Order:
      type: object
      required:
        - merchantId
        - orderNumber
        - items
        - address
        - customer
        - pricing
      properties:
        _id:
          type: string
        merchantId:
          type: string
        orderNumber:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              quantity:
                type: number
              price:
                type: number
        address:
          type: object
          properties:
            ar:
              type: string
            en:
              type: string
            location:
              type: object
        customer:
          type: object
          properties:
            name:
              type: string
            phone:
              type: string
            email:
              type: string
        status:
          type: string
          enum: [pending, confirmed, assigned, picked_up, in_transit, delivered, cancelled, failed]
        courierId:
          type: string
...
    RoutePlan:
      type: object
      required:
        - courierId
        - orders
      properties:
        _id:
          type: string
        courierId:
          type: string
        date:
          type: string
          format: date-time
        status:
          type: string
          enum: [draft, planned, in_progress, completed, cancelled]
        orders:
          type: array
          items:
            type: object
            properties:
              orderId:
                type: string
              sequence:
                type: number
              estimatedArrival:
                type: string
                format: date-time
paths:
...
  /orders:
    get:
      tags:
        - Orders
      summary: List orders
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'

    post:
      tags:
        - Orders
      summary: Create order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '201':
          description: Order created
  /orders/{id}/normalize:
    post:
      tags:
        - Orders
      summary: Normalize order address
      description: Geocode and normalize the order delivery address
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Address normalized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
  /orders/{id}/plan:
    post:
      tags:
        - Orders
      summary: Create route plan for order
      description: Call optimizer service to create optimized route
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                courierId:
                  type: string
                constraints:
                  type: object
      responses:
        '200':
          description: Route plan created
...
  /route-plans:
    get:
      tags:
        - Route Plans
      summary: List route plans
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of route plans
          content:
            application/json:
              schema:
                type: object
                properties:
                  routePlans:
                    type: array
                    items:
                      $ref: '#/components/schemas/RoutePlan'

    post:
      tags:
        - Route Plans
      summary: Create route plan
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoutePlan'
      responses:
        '201':
          description: Route plan created
...